+++ requirements.txt
@@ -1,3 +1,4 @@
+argparse >= 1.4.0
 colorama >= 0.4.3
 simple-term-menu >= 1.2.1
 pyhmy == 0.1.0
