+++ src/menu.py
@@ -1,4 +1,4 @@
-import os, subprocess
+import argparse, os, subprocess
 from os import environ
 from colorama import Fore
 from toolbox.config import EnvironmentVariables
@@ -11,6 +11,7 @@ from toolbox.library import (
     update_text_file,
     first_env_check,
     print_stars,
+    parse_flags
 )
 from toolbox.toolbox import run_full_node, safety_defaults, start_regular_node
 
@@ -26,6 +27,8 @@ if __name__ == "__main__":
         print_stars()
         raise SystemExit(0)
     # clear screen, show logo
+    parser = argparse.ArgumentParser(description='Findora Validator Toolbox - Help Menu')
+    parse_flags(parser)
     loader_intro()
     # check for .env file, if none we have a first timer.
     if os.path.exists(EnvironmentVariables.dotenv_file) is None:
