+++ src/toolbox/toolbox.py
@@ -32,7 +32,8 @@ from toolbox.library import (
     menu_reboot_server,
     finish_node,
     pull_harmony_update,
-    version_checks
+    version_checks,
+    harmony_service_status
 )
 
 def collect_rewards(networkCall):
@@ -405,16 +406,6 @@ def run_regular_node(software_versions) -> None:
                 print_stars()
                 input()
                 start_regular_node()
-
-def harmony_service_status(service = "harmony") -> None:
-    status = subprocess.call(["systemctl", "is-active", "--quiet", service])
-    if status == 0:
-        if service == "harmony":
-            print(f"* {service} Service is:               " + Fore.BLACK + Back.GREEN + "   Online  " + Style.RESET_ALL + Fore.GREEN)
-        else:
-            print(f"* {service} Service is:              " + Fore.BLACK + Back.GREEN + "   Online  " + Style.RESET_ALL + Fore.GREEN)
-    else:
-        print(f"* {service} Service is:               " + Fore.WHITE + Back.RED + "  *** Offline *** " + Style.RESET_ALL + Fore.GREEN)
 
 def service_menu_option() -> None:
     status = os.system("systemctl is-active --quiet harmony")
